Wait for WiFi instead of hanging; update messages
Rather than hanging the application and forcing the user to restart it, the app now asks the user to move within range of a wireless access point and/or enable wireless communications.
It loops until ACU_GetWifiStatus is truthy (ie. the system is connected to WiFi).

Also, error messages have been updated slightly; just basic grammatical and wording tweaks.